Conflicts:
	src/main/resources/hibernate.cfg.xml
	target/classes/hibernate.cfg.xml
	target/classes/org/arQam/AnnonceEfficace/Action/CreateCompteAction.class
	target/classes/org/arQam/AnnonceEfficace/Metier/Utilisateur.class

+++ src/main/java/org/arQam/AnnonceEfficace/Metier/Utilisateur.java
@@ -91,7 +91,7 @@ public class Utilisateur {
   		return null;
   	}
 
-	public Utilisateur load(Long id) {
+	public Utilisateur load(int id) {
    	 SessionFactory sf = HibernateUtil.getSessionFactory();
         Session session = sf.openSession();
      
@@ -129,7 +129,7 @@ public class Utilisateur {
 	public long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 	public String getNomUtilisateur() {
